Created the RecentRecipesResolver, which has a method to get a specified quantity of recent recipes. Added some structure and style for the splash page to display recent recipe cards. SplashComponent has breakPoint observers and waits for recipe data to be resolved.
On the sever side, added a /recipes/recent route which accepts a quantity query param. Added service getRecentRecipes and getRecentRecipeData in the recipes-db. Will retrieve the specified number of recipes, in descending date order.

BUT the LIMIT by quantity is not working. Goes by duplicate recipes, so if there are multiple ingredients, we're not getting the right amount of recipes after they're combined by the recipe-model.